+++ Source/CR4S/Inventory/UI/ItemSlotWidget/BaseItemSlotWidget.cpp
