+++ Source/CR4S/Inventory/UI/ItemSlotWidget/InventoryDummySlotWidget.cpp
