+++ Source/CR4S/Inventory/UI/ItemSlotWidget/InventorySlotWidget.cpp
